Fix ampersand encoding bug and set popup window to be re-used if one is already open.
In this change, I fixed the annoying problem where if a title of a web page
contained an ampersand, it would be double escaped when the bookmarklet
invoked the index.php script in the resulting HTML returned to the
client.

By using document.title instead of the getElementsByTagName, the
html entity for the ampersand will already be decoded before it is
send to the backend.

In this change, I also set the name of the window that opens to allow
the title and description of the task to be changed, to re-use the same
popuip if one is already open. This saves having to close multiple windows
over the process of adding multiple tasks one after another.