Optimize loading of script
To improve the loading responsiveness , moved the script tags to the bottom of the HTML markup,
and changed the order of the Javascript functiosn in code.js to have the helper functions load
first, and the DOM ready event at the bottom.

This should hopefully solve the issue where the UI sometimes froze when the popup is
opened.