Change addTask to use POST instead of GET
Provides better security and allows for much longer task
notes to be submitted when adding a task.

The maximum size of the post body is limited by the PHP
server settings but we should no longer run into the smaller
limit that using GET previously implied.

Requires browser cache to be cleared to load the new
client side Javascript code.