Added star in UI to make task important when adding task
Added a colorful star icon that acts as a checkbox to make a task marked
as important when adding the task.

Upgraded jQuery and jQuery UI to latest versions.
Added fontawesome icon font for the star icon.

Also increased the maximum number of task lists returned to 200
from the previous value of 100.