Improving sorting of task lists with emoji's in the task title
An issue was that if any task list had an emoji assigned to it (via
the MS Todo ability to add an emoji to decorate the list name), it
would be sorted to the bottom of the list of task lists by default
by the default sorting of the MS Todo API.

This change sorts the array of task lists before storing it in the
cache, and before returning it to the front end.

To sort it properly, I remove the starting emoji and any spaces after
it, when doing the sort based on the title. However, the task list
title itself is not changed in the array, as the removal of the emoji
is only done for the purposes of sorting.