Delay updating the items in the dropdown until about 0.5 seconds after the mouse is released
Previously, the dropdown shows the cached task lists first, and then
later updates it with the latest tasks lists. But updating the items
while the dropdown is open and while the user currently has mouse-down
on the dropdown or is actively scrolling is causing problems with the
scroll position being lost when mouse is released.

When the dropdown is updated while the user is actively interacting with
it, the scroll position gets lost. To fix this, I modified the code to
detect when the user is actively interacting with the dropdown and delay
the update until they're done.